+++ cola/core.py
@@ -465,6 +465,17 @@ def _find_executable(executable, path=None):
     return executable
 
 
+def sync():
+    """Force writing of everything to disk. No-op on systems without os.sync()"""
+    if hasattr(os, 'sync'):
+        os.sync()
+
+
+def rename(old, new):
+    """Rename a path. Transform arguments to handle non-ascii file paths"""
+    os.rename(mkpath(old), mkpath(new))
+
+
 if PY2:
     find_executable = wrap(mkpath, _find_executable, decorator=decode)
 else:
@@ -485,6 +496,7 @@ except AttributeError:
 
 realpath = wrap(mkpath, os.path.realpath, decorator=decode)
 relpath = wrap(mkpath, os.path.relpath, decorator=decode)
+remove = wrap(mkpath, os.remove)
 stat = wrap(mkpath, os.stat)
 unlink = wrap(mkpath, os.unlink)
 walk = wrap(mkpath, os.walk)